Fix inconsistent circleci images and failing build
We were building with different versions of node for actual releases
and dev / staging. We should be consistent always so that the artifcacts
we deploy are the same ones we actually tested.

We're also moving from circleci: to cimg: because they deprecated
the old images and asked us to move.

Finally, we're moving to 12.18 because we've seen this image trying to
build on 12.13 without a cache.
We think that builds have only been working because of the circleci
cache of node_modules.
It looks from the builds, like [this
one](https://app.circleci.com/pipelines/github/binary-com/binary-bot/405/workflows/cfe00f01-7ee4-45d5-a4ed-a618e657789f/jobs/426)
that it was getting blockly and gulp from the existing cache

See [this github issue in
blockly](https://github.com/google/blockly/issues/2582).

This is a stepping stone. We need to fix the dependencies, upgrade
blockly, consistently use yarn OR npm, and commit package-lock or
yarn.lock as we work.

This is just an incremental fix.

See also [exiting task](https://redmine.deriv.cloud/issues/8729).